+++ packages/oruga/src/components/popover/examples/delay.vue
@@ -1,0 +1,7 @@
+<template>
+    <section>
+        <o-popover content="The content is shown after 1000ms" :delay="1000">
+            <o-button variant="warning" label="Delayed" />
+        </o-popover>
+    </section>
+</template>