+++ packages/oruga/src/composables/index.ts
@@ -12,3 +12,4 @@ export * from "./useScrollHelper";
 export * from "./useIndexer";
 export * from "./useConfig";
 export * from "./useTrapFocus";
+export * from "./usePopoverAPI";