+++ packages/oruga/src/composables/tests/useEventListener.test.ts
@@ -38,7 +38,7 @@ describe("useEventListener test", () => {
     describe("given event", () => {
         const event = "click";
 
-        test("should add listener", async () => {
+        test("should add listener", () => {
             stop = useEventListener(target, event, listener, {
                 immediate: true,
             });
@@ -93,7 +93,7 @@ describe("useEventListener test", () => {
         test(`should listen event`, async () => {
             useEventListener(target, "click", listener, { immediate: true });
             vi.runAllTimers();
-            target.value!.dispatchEvent(new MouseEvent("click"));
+            target.value?.dispatchEvent(new MouseEvent("click"));
 
             await nextTick();
 
@@ -107,7 +107,7 @@ describe("useEventListener test", () => {
 
             stop();
 
-            target.value!.dispatchEvent(new MouseEvent("click"));
+            target.value?.dispatchEvent(new MouseEvent("click"));
 
             await nextTick();
 
@@ -116,15 +116,15 @@ describe("useEventListener test", () => {
 
         test(`should auto stop listening event`, async () => {
             const scope = effectScope();
-            await scope.run(async () => {
+            scope.run(() => {
                 useEventListener(target, "click", listener, {
                     immediate: true,
                 });
             });
 
-            await scope.stop();
+            scope.stop();
 
-            target.value!.dispatchEvent(new MouseEvent("click"));
+            target.value?.dispatchEvent(new MouseEvent("click"));
 
             await nextTick();